+++ packages/terra-table/src/proptypes/actionShape.js
@@ -1,0 +1,14 @@
+import PropTypes from 'prop-types';
+
+const actionShape = PropTypes.shape({
+  /**
+   * The label for the action button.
+   */
+  label: PropTypes.string.isRequired,
+  /**
+   * The callback for action onClick.
+   */
+  onClick: PropTypes.func.isRequired,
+});
+
+export default actionShape;