+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Column actions example added to `terra-worklist-data-grid`.
+
 ## 1.68.0 - (February 28, 2024)
 
 * Changed