+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Column actions added.
+
 ## 5.8.0 - (February 28, 2024)
 
 * Added